better useClickedEventConnect hook
