+++ applications/fishing-map/src/features/map/use-map-tooltip.ts
@@ -1,6 +1,11 @@
+import { useSelector } from 'react-redux'
 import { Dataview } from '@globalfishingwatch/dataviews-client'
 import { Generators } from '@globalfishingwatch/layer-composer'
-import { InteractionEvent, ExtendedFeature } from '@globalfishingwatch/react-hooks'
+import { InteractionEvent } from '@globalfishingwatch/react-hooks'
+import {
+  selectWorkspaceDataviews,
+  selectTemporalgridDataviews,
+} from 'features/workspace/workspace.selectors'
 
 export type TooltipEventFeature = {
   title: string
@@ -15,29 +20,22 @@ export type TooltipEvent = {
   features: TooltipEventFeature[]
 }
 
-const findDataviewForFeature = (dataviews: Dataview[], feature: ExtendedFeature) => {
-  if (feature.generator === Generators.Type.HeatmapAnimated) {
-    if (!feature.temporalgrid || feature.temporalgrid.sublayerIndex === undefined) {
-      return null
-    }
-    const allTemporalgridDataviews = dataviews.filter(
-      (dataview) => dataview.config.type === Generators.Type.HeatmapAnimated
-    )
-    // TODO We assume here that temporalgrid dataviews appear in the same order as sublayers are set in the generator, ie indices will match feature.temporalgrid.sublayerIndex
-    const dataview = allTemporalgridDataviews[feature.temporalgrid?.sublayerIndex]
-    return dataview
-  } else {
-    return dataviews.find((dataview: Dataview) => dataview.id === feature.generatorId)
-  }
-}
-
-export const getTooltipEvent = (
-  dataviews?: Dataview[],
-  event?: InteractionEvent | null
-): TooltipEvent | null => {
+const useMapTooltip = (event?: InteractionEvent | null) => {
+  const dataviews = useSelector(selectWorkspaceDataviews)
+  const temporalgridDataviews = useSelector(selectTemporalgridDataviews)
   if (!event || !event.features || !dataviews) return null
   const tooltipEventFeatures: TooltipEventFeature[] = event.features.flatMap((feature) => {
-    const dataview = findDataviewForFeature(dataviews, feature)
+    let dataview
+    if (feature.generator === Generators.Type.HeatmapAnimated) {
+      if (!feature.temporalgrid || feature.temporalgrid.sublayerIndex === undefined) {
+        return []
+      }
+
+      // TODO We assume here that temporalgrid dataviews appear in the same order as sublayers are set in the generator, ie indices will match feature.temporalgrid.sublayerIndex
+      dataview = temporalgridDataviews[feature.temporalgrid?.sublayerIndex]
+    } else {
+      dataview = dataviews.find((dataview: Dataview) => dataview.id === feature.generatorId)
+    }
     if (!dataview) return []
     return {
       title: dataview.name || dataview.id.toString(),
@@ -52,11 +50,6 @@ export const getTooltipEvent = (
     longitude: event.longitude,
     features: tooltipEventFeatures,
   }
-}
-
-const useMapTooltip = (dataviews?: Dataview[], interactionEvent?: InteractionEvent | null) => {
-  const tooltipEvent = getTooltipEvent(dataviews, interactionEvent)
-  return tooltipEvent
 }
 
 export default useMapTooltip
